Final push for 0.0.51, going to release dll.

+++ Algebra/Monomial.cs
@@ -77,7 +77,7 @@ namespace Algebra
         {
             get
             {
-                if (moCoefficient < (Int64)0)
+                if (moCoefficient < 0)
                 {
                     return '-';
                 }
